SMBIOS parsing on Linux using mem
