+++ include/osquery/filesystem.h
@@ -195,7 +195,7 @@ Status procReadDescriptor(const std::string& process,
 /**
  * @brief Read bytes from Linux's raw memory.
  *
- * Most Linux kernels include a device node /dev/mem that allows priviledged
+ * Most Linux kernels include a device node /dev/mem that allows privileged
  * users to map or seek/read pages of physical memory.
  * osquery discourages the use of physical memory reads for security and
  * performance reasons and must first try safer methods for data parsing
@@ -206,7 +206,7 @@ Status procReadDescriptor(const std::string& process,
  * This flag/option will cause readRawMemory to forcefully fail.
  *
  * @param base The absolute memory address to read from. This does not need
- * to be page alined, readRawMem will take care of alignment and only
+ * to be page aligned, readRawMem will take care of alignment and only
  * return the requested start address and size.
  * @param length The length of the buffer with a max of 0x10000.
  * @param buffer The output buffer, caller is responsible for resources if
