+++ osquery/tables/specs/x/smbios_tables.table
